feat: 방명록 작성 api
방명록 작성 api 입니다. 수정 조금 했고, 상세조회 api 만들고 있습니다.

+++ back/src/main/java/com/ssafy/rit/back/repository/GuestBookRepository.java
@@ -4,6 +4,6 @@ import com.ssafy.rit.back.entity.GuestBook;
 import com.ssafy.rit.back.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GuestBookRepository extends JpaRepository<GuestBook, Integer> {
+public interface GuestBookRepository extends JpaRepository<GuestBook, Long> {
 
 }
